Merged in feature/CP-336-tk-receipt-delete.4 (pull request #185)
Added exception for data access violations

* Added exception for data access violations

- Added new DataAccessViolationException for situations where a record
exists but it's one that the caller doesn't have access to
- Updated TypedIdSetType so that empty sets are written back to the
database as nulls
- Added new InvalidStateException for cases where the data we've read
back from the database is invalid (e.g. expected one row but got more
than one)
- Added new AccountingDetails to AccountActivity for future work to
send transactions (or account activity in our case) to 3rd party
accounting systems
- Added new CANCELED and CREDIT AccountActivityStatus
- Refactored findByBusinessIdAndReceiptId to be simply findByReceiptId
with code validation of what was read back. It's not super elegant and
we will want to do better
- Moved Table enum to its own class so it could be shared with several
error types

* Merged master into feature/CP-336-tk-receipt-delete.4
* Changed DataAccessViolationException to return 403


Approved-by: Slava Akimov